Finished logarithmic calculation implementation. Code refactoring and build providing.

+++ advanced_calculator/LogarithmicUtilities.h
@@ -4,6 +4,6 @@ int log_calculation_type_input();
 
 long int log_base_input();
 
-long int log_number_input();
+long double log_number_input();
 
 // To Do.
